Fix issues when size_t != unsigned long
This is a patch due to Benjamin Burton. To quote him:

This fixes several compile errors in the GMPXX-related code on systems where
size_t != unsigned long. This happens in particular on 64-bit windows. (You
might not be seeing this because you use MPIR on windows, but regina uses GMP
everywhere so probably I am building this code on additional platforms where
you are not.)

The compile error is related to overloaded functions in GMP that C++ cannot
choose between. I believe the problem is that you are passing size_t arguments
to GMP functions, but these GMP functions do not take size_t arguments. On
normal platforms this is not a problem since size_t == unsigned long, and GMP
functions *do* take unsigned long arguments. However, if size_t != unsigned
long then the compiler needs to cast size_t to a different type that the GMP
functions do accept, and since GMP offers several types to choose from, the
compiler fails because it needs to be told which typecast to choose.

In my patches, I explicitly cast to an unsigned long, which is the largest
native integer type that GMP will accept. This is a cast down (size_t is 8
bytes but long is 4); however, I am (naively?) assuming that these arguments
are indeed sizes of structures (e.g., matrix ranks, vector lengths,
dimensions, row indices, etc.) and so if they ever exceed MAX_LONG then there
are much bigger problems to deal with.

+++ source/libnormaliz/HilbertSeries.cpp
@@ -1120,7 +1120,7 @@ vector<Integer> compute_e_vector(vector<Integer> Q, int dim) {
         }
         E_Vector[i] /= permutations<Integer>(1, i);
         for (j = 1; j < Q.size() - i; j++) {
-            Q[j - 1] = j * Q[j];
+            Q[j - 1] = static_cast<unsigned long>(j) * Q[j];
         }
     }
     return E_Vector;
